some block validation

+++ go/block/extrinsics/disputes.go
@@ -22,7 +22,7 @@ func (d Disputes) ToSumOfValidJudgementsMap() map[[32]byte]int {
 
 type Verdict struct {
 	WorkReportHash [32]byte
-	EpochIndex     uint64 // TODO: must be epoch index of the prior state or 1 less
+	EpochIndex     uint32 // TODO: must be epoch index of the prior state or 1 less
 	Judgements     [constants.NumValidatorSafetyThreshold]Judgement
 }
